+++ src/main/java/de/dm/intellij/LiferayPluginReplacement.java
@@ -1,0 +1,11 @@
+package de.dm.intellij;
+
+import com.intellij.ide.plugins.PluginReplacement;
+
+public class LiferayPluginReplacement extends PluginReplacement {
+
+    protected LiferayPluginReplacement() {
+        super("com.liferay.ide.intellij.plugin");
+    }
+
+}